backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4